Full project restructure into packaged modules
