+++ .cargo/config
@@ -1,4 +1,4 @@
 [alias]
 wasm = "build --release --target wasm32-unknown-unknown"
-unit-test = "test --lib"
+unit-test = "test --lib --features=enable-test-utils"
 schema = "run --example schema"
