+++ src/instantiate/mod.rs
@@ -1,0 +1,1 @@
+pub mod instantiate_contract;
