+++ src/migrate/mod.rs
@@ -1,0 +1,2 @@
+pub mod migrate_contract;
+pub mod version_info;
