+++ src/testutil/mod.rs
@@ -1,0 +1,6 @@
+#[cfg(feature = "enable-test-utils")]
+pub mod instantiation_helpers;
+#[cfg(feature = "enable-test-utils")]
+pub mod test_constants;
+#[cfg(feature = "enable-test-utils")]
+pub mod test_utilities;
