+++ src/testutil/test_constants.rs
@@ -1,0 +1,4 @@
+pub const DEFAULT_INFO_NAME: &str = "admin";
+pub const DEFAULT_CONTRACT_NAME: &str = "wallet.pb";
+pub const DEFAULT_FEE_AMOUNT: u128 = 10000000000;
+pub const DEFAULT_FEE_COLLECTION_ADDRESS: &str = "tp123";
